ai-avatar-interaction/backend/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client (for Ollama)
httpx>=0.27.0

# Speech-to-Text (local Whisper)
openai-whisper>=20231117
torch>=2.0.0
ffmpeg-python>=0.2.0

# Text-to-Speech (Edge TTS)
edge-tts>=6.1.9

# Optional: ElevenLabs (only if enabled)
# elevenlabs>=0.2.27
